Change "flat" delayed eff 0.88 -> 1; fix crash for missing ncap hist
Previously, even if delayed cut were 6 MeV, _relDelEffForBkg would still use
ncap histogram to calculate (unity) bkg rate adjustment. Now we hardcode a
special exception: If delayed-cut is 6 MeV, _relDelEffForBkg just returns 1.
This was needed in order to re-run fit_prep on procut_first, which doesn't have
the ncap histograms. For the sake of consistency, we also modify
actualDelayedEff to return 1 instead of 0.88 when in "flat" mode, since 1 is now
the delayed eff we're using for the background. In principle it shouldn't matter
though, since backgrounds are subtracted (without any delayed efficiency
correction) before the delayed-efficiency correction is applied to the spectrum.